requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
psycopg2-binary
pytest
httpx
pytest-cov